make all editing functionality available on initial import of dataset
